+++ lib/ballercop/logger.rb
@@ -3,22 +3,25 @@ module Ballercop
     LOG_LEVELS = {
       info: 0,
       warning: 1,
-      error: 2,
+      success: 2,
+      error: 3,
     }
 
     def initialize(log_level = nil)
-      @log_level = LOG_LEVELS[log_level || :info]
+      @log_level = LOG_LEVELS[log_level || :error]
     end
 
-    def log(message, log_level = LOG_LEVELS[:info])
-      return unless log_level <= @log_level
-      case log_level
+    def log(message, log_level = :info)
+      return unless LOG_LEVELS[log_level] <= @log_level
+      case LOG_LEVELS[log_level]
       when LOG_LEVELS[:warning]
-        puts "BALLERCOP: #{message}".yellow
+        puts "#{message}".yellow
+      when LOG_LEVELS[:success]
+        puts "#{message}".green
       when LOG_LEVELS[:error]
-        puts "BALLERCOP: #{message}".red
+        puts "#{message}".red
       else
-        puts "BALLERCOP: #{message}"
+        puts "#{message}"
       end
     end
   end