+++ lib/ballercop/patch.rb
@@ -1,31 +1,72 @@
 require 'pry'
+require "stringio"
 
 module Ballercop
   class Patch
-    def initialize(patch, file_path, logger)
+    def initialize(patch, file_path, relative_file_path)
       @patch = patch
       @file_path = file_path
-      @logger = logger || Logger.new(:warning)
+      @relative_file_path = relative_file_path 
     end
     
+    attr_reader :file_path, :fixes_applied, :relative_file_path
+
     def fix
-      offense = fixable_offenses.last
-
-      while offense.present?
-        File.write(@file_path, offense.corrector.rewrite)
-        offense = fixable_offenses.last
+      unless offensive?
+        @fixes_applied = false
+        @no_violations = true
+        return
       end
 
-      unfixable_offenses.each do |each_offense|
-        @logger.log "Can not fix #{each_offense.message} at line #{each_offense.line}", Logger::LOG_LEVELS[:warning]
-      end
+      fix_file if @patch.delta.status == :added
+      fix_patch if @patch.delta.status == :modified
+
+      @fixes_applied = true
+      @no_violations = !offensive?
     end
     
     def offensive?
       offenses.any?
     end
     
+    def result_log(logger)
+      return logger&.log "✅ No errors 🎉 in #{@file_path}", :success if @no_violations
+
+      logger&.log "❌ #{@file_path}", :info
+
+      if @patch.delta.status == :added
+        file_name = @relative_file_path.split('/').last
+        output = []
+        File.foreach(@output_file) { |line| output.append(line) }
+        message = output.compact.reject do |line|
+          line.squish.blank? || (!file_name.include?('tmp') && !line.include?(file_name)) || line.include?("[Corrected]")
+        end
+        message.map { |m| logger&.log "-> #{m.strip}", :warning } if message.present?
+        File.delete(@output_file)
+      end
+
+      if @patch.delta.status == :modified
+        unfixable_offenses.each do |each_offense|
+          logger&.log "-> Can not fix #{each_offense.message} at line #{each_offense.line}", :warning
+        end
+      end
+    end
+
     private
+
+    def fix_file
+      @output_file = "tmp/#{SecureRandom.uuid}.txt"
+      RuboCop::CLI.new.run(['-a', @relative_file_path, "-o#{@output_file}"])
+    end
+    
+    def fix_patch
+      offense = fixable_offenses.last
+
+      while offense.present?
+        File.write(@relative_file_path, offense.corrector.rewrite)
+        offense = fixable_offenses.last
+      end
+    end
 
     def rubocop_registry
       @rubocop_registry ||= RuboCop::Cop::Registry.new(RuboCop::Cop::Cop.all) #TODO: grab cops from .rubocop
@@ -36,17 +77,22 @@ module Ballercop
     end
 
     def rubocop_report
-      config = RuboCop::ConfigStore.new.for(@file_path)
-      processed_source = RuboCop::ProcessedSource.from_file(@file_path, config.target_ruby_version)
-      rubocop_team(config).investigate(processed_source)
+      config = RuboCop::ConfigStore.new.for(@relative_file_path)
+      processed_source = RuboCop::ProcessedSource.from_file(@relative_file_path, config.target_ruby_version)
+      silence_streams(STDERR) do
+        return rubocop_team(config).investigate(processed_source)
+      end
     end
 
-    def offenses
+    def offenses(cache = false)
+      return @offenses if cache && @offenses.present?
       added_lines = patch_added_lines(@patch)
       report = rubocop_report
-      report.offenses.sort
+      @offenses = report.offenses.sort
             .reject(&:disabled?)
             .select {|offense| added_lines.include?(offense.line) }
+      
+      @offenses
     end
 
     def fixable_offenses
@@ -54,7 +100,7 @@ module Ballercop
     end
 
     def unfixable_offenses
-      offenses.select {|offense| offense.corrector == nil || offense.status == :unsupported}
+      offenses(true).select {|offense| offense.corrector == nil || offense.status == :unsupported}
     end
 
     def patch_added_lines(patch)
@@ -64,5 +110,19 @@ module Ballercop
       end
       added_lines
     end
+
+    def silence_streams(*streams)
+      on_hold = streams.collect { |stream| stream.dup }
+      streams.each do |stream|
+        stream.reopen(RUBY_PLATFORM =~ /mswin/ ? 'NUL:' : '/dev/null')
+        stream.sync = true
+      end
+      yield
+    ensure
+      streams.each_with_index do |stream, i|
+        stream.reopen(on_hold[i])
+      end
+    end
+
   end
 end